pyproject.toml
[project]
name = "IASEB"
version = "1.0.0"
description = "Interaction-Aware Spatio-temporal Entity Benchmark for Video Grounding"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Alejandro Parcedo"},
]
keywords = ["video grounding", "spatio-temporal", "benchmark", "vision-language models"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
]
dependencies = [
    "torch>=2.0",
    "transformers>=4.40",
    "opencv-python>=4.8",
    "numpy>=1.24",
    "Pillow>=9.0",
    "tqdm>=4.60",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
analysis = [
    "pandas>=2.0",
    "scipy>=1.10",
    "matplotlib>=3.7",
]
classification = [
    "openai>=1.0",
    "python-dotenv>=1.0",
]
qwen = [
    "qwen-vl-utils>=0.0.8",
]
all = [
    "IASEB[analysis,classification,qwen]",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["IASEB"]
